% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwindow.R
\name{open_window}
\alias{open_window}
\title{Open Window}
\usage{
open_window()
}
\arguments{
\item{rotate}{Default 0. Rotates the output. Possible values: 0, 90, 180, 270.}
}
\description{
Open Window
}
\examples{
#internal
}
