% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwindow.R
\name{open_window_image}
\alias{open_window_image}
\title{Open Window Image}
\usage{
open_window_image(image, width = 640, height = 360)
}
\arguments{
\item{width}{Window width}

\item{height}{Window height}
}
\description{
Open Window Image
}
\examples{
#internal
}
\keyword{internal}
